\name{biomasseX}
\alias{biomasseX}
\docType{data}
\title{A factorial input design for the main example}
\description{
 Factorial design (resolution V) data for the 7 WWDM model input factors
}
\usage{data(biomasseX)}
\format{
  A data frame with 2187 observations on the following 7 variables.
  \describe{
    \item{\code{Eb}}{First WWDM input factor name}
    \item{\code{Eimax}}{Second WWDM input factor name}
    \item{\code{K}}{Thirth WWDM input factor name}
    \item{\code{Lmax}}{Fourth WWDM input factor name}
    \item{\code{A}}{Fifth WWDM input factor name}
    \item{\code{B}}{Sixth WWDM input factor name}
    \item{\code{TI}}{Seventh WWDM input factor name}
  }
}
\examples{
data(biomasseX)
## maybe str(biomasseX) ; plot(biomasseX) ...
}
\keyword{datasets}
\seealso{\code{\link{biomasse}}, \code{\link{biomasseY}}}
