\name{biomasseY}
\alias{biomasseY}
\docType{data}
\title{Output of the biomasse model for the plan provided in the package}
\description{
 Simplified output of the biomasse model (one column per
 decade), especially generated for examples in the package help files
}
\usage{data(biomasseY)}
\format{
  A data frame with 2187 rows and 22 output variables (one per decade).
}
\examples{
data(biomasseY)
dim(biomasseY)
}
\keyword{datasets}
\seealso{\code{\link{biomasse}}, \code{\link{biomasseX}}}
