\name{Climat}
\alias{Climat}
\docType{data}
\title{Climate data}
\description{
  Climate data for the WWDM model (needed by the \code{biomasse} function)
}
\usage{data(Climat)}
\format{
  A data frame with 3126 observations on the following 4 variables.
  \describe{
    \item{\code{ANNEE}}{a factor with levels 1 to 14, indicating 14
      different years}
    \item{\code{RG}}{daily radiation variable}
    \item{\code{Tmin}}{daily maximum temperature}
    \item{\code{Tmax}}{daily minimum temperature}
  }
}
\source{
Makowski, D., Jeuffroy, M.-H., Gu\'{e}rif, M., 2004 Bayesian methods for
 updating crop model predictions, applications for predicting biomass and
 grain protein content. In: Bayesian Statistics and Quality Modelling in
 the Agro-Food Production Chain (van Boeakel et al. eds), pp. 57-68.
 Kluwer, Dordrecht.

Monod, H., Naud, C., Makowski, D., 2006 Uncertainty and sensitivity
 analysis for crop models. In: Working with Dynamic Crop Models (Wallach
 D., Makowski D. and Jones J. eds), pp. 55-100. Elsevier, Amsterdam
}
\keyword{datasets}
