\name{grpe.gsi}
\alias{grpe.gsi}
\title{Group factor GSI, obsolete function}
\description{
 An obsolete function that computed the GSI of a group factor as one factor
}
\usage{
grpe.gsi(GSI, fact.interet)
}
\arguments{
  \item{GSI}{a gsi or dynsi object}
  \item{fact.interet}{input factor to be grouped}
}
\note{This is essentially an internal function for the \pkg{multisensi} package}
\author{M. Lamboni}
\keyword{internal function}
