\name{print.dynsi}
\alias{print.dynsi}
\title{print DYNSI }
\description{
  A function to print DYNSI results
}
\usage{
\method{print}{dynsi}(x, ...)
}
\arguments{
  \item{x}{a dynsi object}
  \item{...}{print parameters}
}
\seealso{\code{\link{dynsi}}, \code{\link{multisensi}}}
\author{M. Lamboni}
