\name{summary.dynsi}
\alias{summary.dynsi}
\title{dynsi summary}
\description{
  Function to summarize the dynamic sensitivity results
}
\usage{
\method{summary}{dynsi}(object, ...)
}
\arguments{
  \item{object}{a dynsi object}
  \item{...}{summary parameters}
}
\seealso{\code{\link{dynsi}}}
\author{M. LAMBONI}
