\name{summary.gsi}
\alias{summary.gsi}
\title{summary of GSI results}
\description{
  function to summarize the GSI results
}
\usage{\method{summary}{gsi}(object, ...)}
\arguments{
  \item{object}{a GSI object}
  \item{...}{summary parameters}
}
\seealso{\code{\link{gsi}}, \code{\link{multisensi}}}
%\author{C. Bidot, M. Lamboni}
